feat: add postinstall hook
